Rename files to camel case
