+++ src/AccordionItemBody/AccordionItemBody.spec.js
@@ -5,7 +5,7 @@ import { mount } from 'enzyme';
 import { Provider } from 'unstated';
 import ItemContainer, { resetNextUuid } from '../ItemContainer/ItemContainer';
 import AccordionContainer from '../AccordionContainer/AccordionContainer';
-import AccordionItemBody from './accordion-item-body-wrapper';
+import AccordionItemBody from './AccordionItemBody.wrapper';
 
 describe('AccordionItemBody', () => {
     let accordionStore;
